Initial Jupiter dimension

+++ src/main/java/com/shim/celestialexploration/entity/mob/slimes/SulfurCube.java
@@ -39,7 +39,7 @@ public class SulfurCube extends MagmaCube {
         if (fluidTagKey == FluidRegistry.SULFUR_TAG) {
             CelestialExploration.LOGGER.debug("SulfurCube in Sulfur");
             Vec3 vec3 = this.getDeltaMovement();
-            this.setDeltaMovement(vec3.x, (double)(0.22F + (float)this.getSize() * 0.05F), vec3.z);
+            this.setDeltaMovement(vec3.x, 0.22F + (float)this.getSize() * 0.05F, vec3.z);
             this.hasImpulse = true;
         } else {
             super.jumpInLiquid(fluidTagKey);
